third_party/blink/web_tests/external/wpt/html/cross-origin-opener-policy/coep-navigate-popup.https-expected.txt

This is a testharness.js-based test.
PASS Popup navigating to same-origin with coop/coep
FAIL Popup navigating to same-site with coop/coep assert_equals: expected "" but got "Popup-navigating-to-same-site-with-coop/coep"
FAIL Popup navigating to same-origin with no coop/coep assert_equals: expected "" but got "Popup-navigating-to-same-origin-with-no-coop/coep"
FAIL Popup navigating to same-site with no coop/coep assert_equals: expected "" but got "Popup-navigating-to-same-site-with-no-coop/coep"
FAIL Popup navigating to same-origin with coop/no coep assert_equals: expected "" but got "Popup-navigating-to-same-origin-with-coop/no-coep"
FAIL Popup navigating to same-site with coop/no coep assert_equals: expected "" but got "Popup-navigating-to-same-site-with-coop/no-coep"
FAIL Popup navigating to same-origin with no coop/no coep assert_equals: expected "" but got "Popup-navigating-to-same-origin-with-no-coop/no-coep"
FAIL Popup navigating to same-site with no coop/no coep assert_equals: expected "" but got "Popup-navigating-to-same-site-with-no-coop/no-coep"
PASS Popup navigating to same-origin with coop unsafe-inherit/coep
FAIL Popup navigating to same-site with coop unsafe-inherit/coep assert_equals: expected "" but got "Popup-navigating-to-same-site-with-coop-unsafe-inherit/coep"
FAIL Popup navigating to same-origin with coop unsafe-inherit/no coep assert_equals: expected "" but got "Popup-navigating-to-same-origin-with-coop-unsafe-inherit/no-coep"
FAIL Popup navigating to same-site with coop unsafe-inherit/no coep assert_equals: expected "" but got "Popup-navigating-to-same-site-with-coop-unsafe-inherit/no-coep"
Harness: the test ran to completion.
